more bug fix
