+++ js/handlers/index.mjs
@@ -2,8 +2,8 @@ import { toggleMobileNav } from "./mobile-nav-toggle/index.mjs";
 import { toggleComments } from "./posts/comments-handler.mjs";
 import { toggleMediaBtn } from "./posts/media-btn-handler.mjs";
 import { toggleTagBtn } from "./posts/tag-btn-handler.mjs";
-import { loginAccount } from "./registration/login-form-handler.mjs";
 import { logoutAnchorAccount } from "./registration/logout-anchor-handler.mjs";
+import { loginFormHandler } from "./registration/login-form-handler.mjs";
 import { registerFormHandler } from "./registration/register-form-handler.mjs";
 
 export function postHandlers() {
@@ -16,5 +16,5 @@ export function postHandlers() {
 
 export function signupHandlers() {
    registerFormHandler();
-   loginAccount();
+   loginFormHandler();
 }
