+++ js/handlers/registration/login-form-handler.mjs
@@ -1,19 +1,9 @@
-import { authenticateProfile } from "../../api/auth/authenticate-profile.mjs";
+import { loginAccount } from "../../api/account/login-account.mjs";
 
-export function loginAccount() {
+export function loginFormHandler() {
    const loginForm = document.querySelector("#form-to-login");
-
    loginForm.addEventListener("submit", (event) => {
       event.preventDefault();
-      const formElements = event.target;
-      const loginCredentials = {
-         email: formElements.email.value,
-         password: formElements.password.value,
-      };
-      const method = formElements.method;
-      const action = formElements.attributes.action.value;
-
-      console.log(loginCredentials, method, action);
-      authenticateProfile(loginCredentials, method, action);
+      loginAccount(event);
    });
 }
